refactor: create useThemeExtractor hook
