+++ packages/vscode-extension/src/network/components/NetworkLogDetails.tsx
@@ -13,7 +13,8 @@ import TimingTab from "./Tabs/TimingTab";
 import { useNetwork } from "../providers/NetworkProvider";
 import { NetworkLog } from "../types/networkLog";
 import { ResponseBodyData } from "../types/network";
-import { ThemeData, ThemeType } from "../types/theme";
+import { ThemeData } from "../types/theme";
+import useThemeExtractor from "../hooks/useThemeExtractor";
 
 const VSCODE_TABS_HEADER_HEIGHT = 30;
 
@@ -37,61 +38,12 @@ interface Tab {
 }
 
 const NetworkLogDetails = ({ networkLog, handleClose, parentHeight }: NetworkLogDetailsProps) => {
-  const { getResponseBody } = useNetwork();
-  const [editorThemeData, setEditorThemeData] = useState<ThemeData | undefined>(undefined);
   const [responseBodyData, setResponseBodyData] = useState<ResponseBodyData | undefined>(undefined);
-
   const { wasTruncated = false } = responseBodyData || {};
 
-  useEffect(() => {
+  const { getResponseBody } = useNetwork();
 
-
-    const body = document.querySelector("body");
-    if (!body) {
-      return;
-    }
-
-    const themeTypeAttibute = "data-vscode-theme-kind";
-    const themeIdAttribute = "data-vscode-theme-id";
-
-    const classObserver = new MutationObserver((mutations) => {
-      mutations.forEach((mut) => {
-        if (
-          mut.type !== "attributes" &&
-          mut.attributeName !== themeTypeAttibute &&
-          mut.attributeName !== themeIdAttribute
-        ) {
-          return;
-        }
-
-        const newThemeData = {
-          themeType: body.getAttribute(themeTypeAttibute) as ThemeType,
-          themeName: body.getAttribute(themeIdAttribute) || "",
-        };
-
-        setEditorThemeData(prev => {
-          if (
-            prev?.themeName === newThemeData.themeName &&
-            prev?.themeType === newThemeData.themeType
-          ) {
-            return prev;
-          }
-          return newThemeData;
-        });
-      });
-    });
-
-    classObserver.observe(body, { attributes: true });
-
-    const initialThemeData: ThemeData = {
-      themeType: body.getAttribute(themeTypeAttibute) as ThemeData["themeType"],
-      themeName: body.getAttribute(themeIdAttribute) || "",
-    };
-
-    setEditorThemeData(initialThemeData);
-
-    return () => classObserver.disconnect();
-  }, []);
+  const editorThemeData = useThemeExtractor();
 
   useEffect(() => {
     getResponseBody(networkLog).then((data) => {
@@ -107,7 +59,7 @@ const NetworkLogDetails = ({ networkLog, handleClose, parentHeight }: NetworkLog
     {
       title: "Payload",
       Tab: PayloadTab,
-      props: { editorThemeData }
+      props: { editorThemeData },
     },
     {
       title: "Response",
@@ -120,7 +72,7 @@ const NetworkLogDetails = ({ networkLog, handleClose, parentHeight }: NetworkLog
       Tab: TimingTab,
     },
   ];
-  
+
   return (
     <>
       {/* TODO: use VscodeToolbarButton when it will be available in @vscode-elements/react-elements  */}
@@ -132,7 +84,7 @@ const NetworkLogDetails = ({ networkLog, handleClose, parentHeight }: NetworkLog
           <Fragment key={title}>
             <VscodeTabHeader className="network-log-details-tab-header">
               <div>
-                {<span className="mtk17">{title}</span>}
+                {title}
                 {warning && <span className="codicon codicon-warning" />}
               </div>
             </VscodeTabHeader>
